lists: simplify checks with minimal version of updater check
There has to be step between 3.x and 4.x version because of other
reasons than updater's incompatibility. That gives us possibility to
define new lowest supported version of updater-ng for 4.x releases.
Thanks to that we can remove some obsolete crap.

It is planned to bump minimal version of updater even higher as soon as
relevant updater's release is out.