compile_fw: fix git configuration and that way clean builds
Previous change to compile_fw broken execution in clean directory
(without git). compile_fw could not been run, not even to display help.

This commit changes how configuration is applied to git tree of OpenWRT.
It is used only when needed and never set to git tree it self.

This introduces new _git function. This function is not required to be
used every time but for `git commit` and `git am` it is necessary to use
it instead of plain git.