compile_pkgs: simplify configure command and few other things
In configure command there are two major changes. VERSION_REPO is now
set to new path path that is unfortunately board unspecific but as
only one makes sense in new repository arrangement. Second change is
that we no longer build image builder and SDK. Image builder does not
make sense for use. We use updater to build images (medkits). SDK is at
current state useless. It does not contain anything other then compiler,
libraries are missing.

Second change is primarily done because of next change that is going to
be splitting off compile command to few subcommands and making sure that
we build only packages not anything else.