Revert "Add hack to not ignore dependency on missing packages"
This reverts commit 69b6fe0d514b55ab1d2141a54a79fd60433c88b1.

It broke horribly as expected :-) Trick is that if package is not found, the
dependency is mallformed anyway, so it is better to just get rid of it and hope
for the best than to break with weird error. Correct solution would be exiting
with the correct error but first of all make dependencies accross feeds work.