Nonabsolute lists
This creates lists that can be included in to each other. This is done
by hack that original script calling one of these lists passes base URI
to access sources and that is later reused. This is how relative URI
should work in future but it is not completed now so we can overcome it
this way and have almost same support.

This also means that we no longer have to specify target branch for
lists because it no longer matters.

Fallback branch is for now dropped because we are not doing small build
at the moment. We should in future return to them but they will be
implemented differently then fallback branch is.

This also does some base list cleanup.