Revert "Try to fix python packaging"
This reverts commit d38d7a1e58a40946cdf43455342ccf124ef3a21e.

We foundout why these weird dependencies are there. That is because
dependencies are resolved for whole package not just for specified build
package. This results to build of all python versions even if only one
of them in reality is needed. These dependencies mask that and ensure
that if package is going to be build that its python version is skipped.