Revert "Try to fix python packages"
This reverts commit 99cc9e6db9b780bd49bf7c194ed3a8dd8c8f2565.

We foundout why these weird dependencies are there. That is because
dependencies are resolved for whole package not just for specified build
package. This results to build of all python versions even if only one
of them in reality is needed. These dependencies mask that and ensure
that if package is going to be build that its python version is skipped.