compile_pkgs: fix problem with ignored packages from feeds
This fixes problem with ignored packages that were introduced by patch.
What this does is that it installs packages after patching is done. It
also first updates indexes again before installation.

Because this feature was originally split between two function but
patch_feeds had to be inserted to the middle of get_feeds I decided to
drop both and used upper repatch_feeds instead. This means that we are
no longer able to get and patch feeds separately but I think that this
separation was almost useless.