generate_medkit: include test key if branch specified
The idea is that if you set some branch then it is most probably not
deploy signed branch and so we should include test key. On the other
hand if it is a deploy branch then this is not huge problem. The key
would be removed with first updater run.