lists: Report target Turris OS
This should help us to debug system configuration better to see what
exactly was targeted version. This also helps users to see what they are
dealing with. It in some sense informs them if they are upgrading to new
version or if that is just packages change (because of configuration
change) in same release.