compile_pkgs: don't run it under sudo and build everything
Parameter -e is now obsolete and by default, we want to build everything.
Removed it from turris-build.conf.example

Because of that we or users don't need to select packages in make
menuconfig which we or they want to build. Now each package will be selected
but not compiled which depends on user choice.

Also, compile_pkgs shouldn't be run under root or with sudo. Do it as
normal user.