compile_pkgs: rename secondary options section
Configuration override options was misleading. In reality those are
options you most probably don't want to mess too much with so let's call
them rather advanced options.

This also moves PUBLISH_BRANCH in example configuration file to
section: don't mess with. This is because now we provide it from
defaults and user just should not mess with it unless he know what he is
doing.