lists: fix problems with not removed luci applications
This reverts to previous approach for Luci localization packages. It
introduces back problems that language packages are installed only to
apps introduced by updater lists but there is no way around it. This
implementation was problematic as it kept installed packages in place
because of language packages depending on package that was suppose to be
removed. This meant that such package was hold in place by language
package that was in turn hold in place by information that parent
package is installed. It created cycle that updater was not able to
break (only opkg). This fixes this issue.