lists/pkglists: use new pkglists options
This allows more granular control for big pkglists.

The consideration was to either introduce dependencies between lists or
one level of additional options where dependency implies that lower
layer required upper layer. The second option is easier to implement so
we went with it. Every pkglist now can have options that are passed to
it from updater script on router and allows various options to be
disabled and enabled.

Migration to this setup is ensured by fix package we include when older
version of pkglists is installed.