lists: mark wget-nossl as unwanted alternative
OPKG seems to be using uclient-fetch if wget is not available but once
wget command is on system it uses it. The problem is that we have URLs
with SSL and so we need wget with SSL support but in default updater
minimalizes amount of installed packages so it chooses wget-nossl
variant over full wget.