patches/luci: drop firmware flash page
This removes link to flash Luci page from menu. The page is still
accessible but now user no longer has direct access to that page.

This is because we do not support updates or backups provided by that
page. Its behavior is potentially harmful and misses script for firmware
upgrade to work correctly. Removing it from menu is short patch that can
be applied for long time and should prevent users from tinkering with
something that is not intended to work on Turris.