compile_pkgs: rework how we detect remote hash
This does the same thing as git status but is safe for script usage. It
just goes from tip (HEAD) trough all commit hashes and requests git to
name them using symbolic link. We limit that to only references from
origin so default remote references. This is to get commit that is in
upstream instead of commit hash of tip that is patch.