patches/luci: report remote commit hash instead of tip
This should help with debugging. The hash used previously was to commit
that is created by applying patch and it was implosible to found
reference for sources because of that. This instead looks for first
original commit and uses that instead.

Packages are reinstalled non the less because of build time.