Revert "patches/openwrt: try to drop kernel magic"
This reverts commit 25d090d9ed68f91296255c52c49a99f482077a0c.

This is required because kmods are not correctly updated. This is
because when kmod is check if it should be upgraded the check works only
with version and because packages has same version localy as on server
the upgrade is skipped. We need kmod to change its version so it would
be reinstalled.