compile_pkgs: drop kernel config push
This was colliding with OpenWrt's own configuration push and was kind of
hack. Replacement is to use OpenWrt's approach. That is same but in
OpenWrt configuration. To ensure that configuration is used by OpenWrt
KConfig in OWrt tree has to be added.
The new approach also simplifies some stuff because we can let KConfig
to do configuration for us. There is no need to specify specific
options to configuration in this repository. An example can be Turris
Omnia LEDS support (file target/linux/mvebu/image/Config.in):

config KERNEL_LEDS_OMNIA
	bool "Build Turris Omnia LEDS support to kernel"
	depends TARGET_mvebu
	default y if TARGET_mvebu_cortexa9_DEVICE_cznic_turris-omnia
	default n

This adds option CONFIG_KERNEL_LEDS_OMNIA that is copied to kernel
configuration as CONFIG_LEDS_OMNIA. It is available only for
TARGET_mvebu as it is part of patch for that target. In default it is
set to no. No is copied to kernel configuration as "# CONFIG_LEDS_OMNIA
is not set" and that ensures that for other boards kernel compiles as it
knows how to set this option. The default for Turris Omnia board is yes
and that inserts correct "CONFIG_LEDS_OMNIA=y" to kernel config.