patches/packages: mariadb: remove patch for broken priority of feeds
My patch, which was present in the hack folder, is no longer applying as
mariadb was updated in OpenWrt 18.06, 19.07 and master branches.
This patch was meant to be a workaround as something in OpenWrt was
changed and it caused inconsistency in the installation of packages and we
would like to have provided mariadb in our feed, this was the easiest
and a quicker solution at that time.

Solution is provided in this commit: a42ef21a740653ee76146bfccee1ab6afbd65539