compile_pkgs: use flock to not colide in git mirrors
This locks git mirror directory when we are updating mirrors. This
should solve race condition with multiple builds failing to update
mirrors and failing that way completely.

Only update is protected. The expectation is that initialization is
happenning only once and is protected by non-existence of directory. If
there is going to be problem with initialization as well then we can add
locking as well.