lists/base-fix: switch to finished replan
It makes sense to replan immediately here but this kind of replan
possibly breaks stuff so we should rather not use it unless it solves
some problem that can't be solved any other way.

We can revert this if we found out that we really need immediate replan
but current effects are that it possibly would just temporally remove
some of the packages that were moved to option condition. That should
not be the end of the world.