lists/foris: handle all Foris language packages centrally
This has various advantages such as that we do not have to care to
always set appropriate requests for language packages as well as for
core package. It also ensures that localization is installed even if
package is installed outside of package list it is mentioned in.

There is also improvement for reForis. Now when you install reForis its
plugins are going to be managed relative to those requested to be
installed for Foris.