compile_pkgs: verify generated config against our preferences
This checks generated config after Kconfig pass if it still contains
preferences we have in files in configs directory.

This also drops some of the options in our config that are obsolete or
invalid.

CONFIG_VERSION_REPO is overwritten by setting pushed dynamically in
config by compile_pkgs so it makes no sense to have it here.

CONFIG_BUSYBOX_DEFAULT_CROND seems to be impossible to change and is
also default value that we later override with
CONFIG_BUSYBOX_CONFIG_CROND.

All PACKAGE options are not necessary. They are all enabled to 'm' but
getting them to 'y' would require a lot of works with gain that build
would in some cases failed instead of passed without that package. Not
worth it.

CONFIG_DEFAULT_TARGET_OPTIMIZATION is not applies so it is dropped here
without replacement. Same as one occurrence of CONFIG_CPU_TYPE.