lists/luci: handle all luci-i18n-* packages centrally
These packages are same as language packages for Foris

There is one speciality and that is that some of Luci's packages have
English localization package as well so we have to add "en" to
languages.

There is full list of all Luci applications there are in repository. We
should update this in future or modify the updater functionality so it
would be possible to check for existence of package in repository.