lists/base-fix: add fix for updater v65.0 Alternatives
This fixes problem that updater running is unable to handle alternatives
but with installing new one the script that was handling it previously
was removed. This installs one-time fix package that runs that original
script on its removal. The reason why on its removal is that it lets
first install all packages by original updater and new updater later
just removes this fix package and in that time it makes sense to
consitently update alternatives the same way as we did it previously
with postupdate hook script.