kmod.list: Install support for USB Mass Storage devices by default
We have Storage plugin in Foris installed by default, but it is not
possible to use it without an installed NAS package list, which contains a
lot of packages. Let's move related support for USB Mass Storage devices
to kmod.list, which will be installed by default for all installations.