lists: keep fix for alternatives in system for now
This should fix any inconsistencies in system done by opkg as well as
updater.

It turns out that we really have to handle busybox links without
Alternatives as soon as possible. Without it some packages fail to work
if busybox is updated or reinstalled and alternatives are overwritten by
busybox links.