patches/openwrt: add autoconf-archive tool
This tool is required for updater. In reality it is set of scripts that
are extension for autoconf. Most of the packages do not need it because
even if they use it they use distribution archives that contain copy of
these scripts. We could do the same although for development purposes it
is easier to just include those scripts in OpenWrt build system over
needing to generate distribution archives.