hardening: disable SECCOMP
Let's disable SECCOMP.
Seccomp and jail are used and developed just by one person in OpenWrt.

Services like transmission do not start (there are missing syscalls!), altougth if there are
added missing syscalls transmission does not start through procd.
There are no details how to proceed further without documentation. Some processes are jailed
and we can not check which one each time (and if it is backported to
OpenWrt 19.07) and fix it.

It is not enabled by default in OpenWrt and it has several reasons for
it. For example, umdns service ends up in crash loop.

We can not delay Turris OS 5.0 anymore as seccomp adds us overhead
for now.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>