lists: install python3-python if there is no python-base
Package python-base was providing python link to python2. By removing
python-base (and effectively this way Python2) from system this removes
this link as well. This link is commonly expected to be present so we
add it for python3 if there is not python-base.

This condition also as corner case support case when Python3 is not
installed.