configs/common: allow setting hostapd debug level to highest level
This setting limits amount of log messages available in hostapd. Setting
this to zero allows us to run hostapd in debug mode.

This has effect only if hostapd is started with arguments enabling debug
(specifically -d). In default this changes nothing. It only allows
debugging. Manual start of hostapd is required to add these arguments or
netifd scripts have to be hacked.