patches/openwrt: iwinfo: add device ID for Marvell 88W8997 and Atheros AR9287
Pali sent these two patches to mailing list openwrt-devel to almost 2 weeks ago and haven't received a response, yet.
Most likely those are not getting backported to OpenWrt 19.07 and we
would like to have in our Turris OS distribution, which is based on top
of it.

This fixes issue that some of ours Wi-Fi cards included in our units
were recognized (e.g. in LuCI) as Generic Wi-Fi card.

I need to rebase one of his patches to apply cleanly to it.