lists/migrate3x: drop old and broken Turris 1.x redirect to HBD
There were times when only supported branch for Turris 1.x routers was
HBD. Becasue of that repository used in migration was HBD in case of
Turris 1.x. With changes merged to HBL and HBD this is broken and also
wrong. We have some support in HBK so it should point to HBK to be
consistent between routers.