lists/base-fix: Reset u-boot environment configuration
This triggers removal of uboot environment configuration from system so
we can set there correct one by installing new version of uboot-tools.

In case of Turris Mox we have to do also additional trickery as file we
want to generate is part of mox-support package originally. With this
update it is dropped but we have to let it be removed fist before we
recreate it by installing new version of uboot-tools otherwise updater
would remove that file as part of mox-support package cleanup phase.