lists/migrate3x: change root package for replan to tos3to4-early
This solves issue that tos3to4-early depends on switch-branch and that
one depends on updater-supervisor and that one depends on updater-ng.
In the end that creates with additional dependency of updater-ng on
tos3to4-early cycle dependency and updater won't allow that with
critical package.

This just uses tos3to4-early for same purpose as previously updater-ng.
It is just rude fix to break cycle.