patches/openwrt: Add link signature to packages
This should in future replace potentially automatic package version
bumping patch. This should be cleaner solution for same problem as it
won't tamper with versions and does not need server-side database.

Problem with tampering with version is that matching version exactly was
then pretty much broken. Once package was bumped the version did not
match. It is impossible to update dependency easily enough when package
bump occurs.

The idea is that updater just matches signature that should be
consistent between builds (unless dynamic link dependencies changed of
course). When build results to package with different link dependencies
then signature changes and updater reinstall such package to version
from repository that is consistent. Version stays the same all the time.

Signature is generated from two sources. Dynamic library names and
dynamic symbols that are linked to those libraries are included.