compile_pkgs: split force-make to separate script
Running make multiple times for builds with last debug fallback run is
handy not even for calls available trough compile_pkgs. For example
building single selected package is not possible using compile_pkgs and
that script is not intended to even do it. The clean solution because of
that is to just simply extract logic for forced make run to separate
script that can be used without compile_pkgs,