feeds/turrispackages: nextcloud: remove php7-mod-hash
Turris packages feed is using master for fixup releases and develop for
minor/major releases, but in case of different OpenWrt versions, there
might be removed or changed name of some packages.

I found that php7-mod-hash wont be present in OpenWrt 20.xx release
as there are using php 7.4.x and it is part of it. On the other hand,
there is php7.2.x in OpenWrt 19.07.