lists/bootstrap: fix invalid usage of environment variables
This fixes bug that caused normal medkits to not be generated as
contract with no name was not located.

The reason is that Lua's string when they are empty are still true
(contrary to other languages such as Python). This adds explicit checks
if they are non-empty as that is more common and more expected.

This also fixes potential problem when some environment variable is not
defined. This does no happen with our script but can happen if that
script is not used.