lists: fix increased priority from l10n packages
Standard for_l10n function provided by root updater script is creating
requests with default priority 50. This effectively increases priority
of all packages using it to include their languages.

The solution here is to decrease priority for language packages to
something very small (10 was chosen).

Also because we had to have copy of for_l10n function in bootstrap it
was instead moved from there to utils and provided as our own version
instead of updater's one.
Original function can potentially be dropped later on as Turris lists
are no longer using it thanks to this change.