lists: add USB and PCI drivers
These contain database of USB and PCI device mapping to
drivers/packages. Idea is that updater with additional Lua script on
router can install this way drives automatically depending on what is
actually connected to router.

It also allows installtion of all or just some group of drivers. That is
to include them optionally also in medkit.