generate_medkits: this is more general script instead of generate_medkit
We need to generate more than one medkit per target. That can be
controlled by CI but it is better to have it all at once controlled by
turris-build repository. It also partially hides all possible options
generate_medkit provides. Most of them are to have versatile tool to
generate medkit but are not intended to be used in CI to build real
medkits. With this script (named generate_medkits to be consistent with
generate_lists) only minimal required info is required from user and
rest is provided depending on what kind of medkit is build.

To have repository clean generate_medkit was moved to helpers directory.

There are hints on future variants of medkits. At the moment primary use
is for contracts but we can also provide minimal medkits (pretty much
compared to what we have), beef up default medkit (just include some
more common package lists) and also for example have full medkit with
all package lists and drivers and languages (it is question if we even
want that). It is also expected to have more targeted medkits such as
broadband variant that would out of the box provided support for LTE
modems and so on.