helpers/new_release: fix invalid match for modification date
Fields should be compared case insensitive as they can be in any such
format. Using awk instead to compare and correctly parse HTTP head
format.

There is known problem with this code and that is it can't parse
date-time formats with colons. The reason is that printing original
string ($0) in awk sneaks in to it some unwanted characters such as \r
and probably others that are not considered as whitespaces by date
command. This way we take only text.