lists: refactor and make Foris not the top level package
There are multiple changes to updater's lists. In most cases it is about
moving packages around and grouping them together to simplify future
maintenance.

Significant change is usage of conditional requests. Updater now
installs Foris plugins not because some package list was selected but
because some package was installed. This makes it so users running
minimal setup (without web interface) can use package lists and still
have full system with it.

One important change to note here is that SSH client was moved from
minimal base list to terminal-apps. The reason is that it is not
essentially required for router's functionality or access. It is more
about secondary usage from CLI so it should have been part of
terminal-apps all time along.