patches/openwrt: disable root account
This disables root account.

Placing of this patch expects it to be caried for a long time so it was
moved to second possition while current second patch was moved to thid.
That makes it of same number as existing patch but that patch is going
to be dropped in future so no harm to have them with same number for
now.