lists: move to separate repository and improve scripts
There are two changes in this commit. First most noticable is removal of
all updater lists.

The second but more significant change is pretty much reformat of all
generate scripts. It extracts common git and configuration handling from
compile_pkgs to generate_common.
The advantage of this is of course code sharing but also that other
generate scripts now get features such as local git mirror.