lists/base-fix: reinstall luci-base to prevent files collision
Merge of luci-lighttpd and patch that was providing webapps files for
luci to package turris-webapps-luci caused file collision. The cause is
that patch adding webapps did not change version of luci-base package.
The solution is to just reinstall luci-base package to remove
conflicting files.
This affects mostly only HBL and HBD. It is most probable that update
between versions is going to bump LuCI version and solve it that way
insteand. At the same time it won't harm to have this.