gitlab-ci: move Dockerfile to hidden directory
Dockerfile does not essentially have to be in root or repository. It is
Gitlab CI specific non the less.
This has one slight improvement and that is it is build faster as docker
does not have to pull in whole repository first.
To build image one additional step is to just navigate to .gitlab-ci
directory before docker build call.