.gitlab.ci: refactor package lists and add verify stage
There was an added test for package lists, but it runs twice.
If you check in the previous changes then there are two jobs, which
do practically the same. In generate pkglists there are generated
lists with debug parameter, but we don't work with that anymore, which
is wrong. And because of that, we had another job check generated
pkglists, where we are generating lists once again and then we are
checking if those have been generated.

Solution: Don't waste hardware resources and run it just once. When package lists
are generated then the artifact is uploaded and uses in the next job in
the next stage