GitLab CI: add generate and check medkit
We are testing in Gitlab CI if patches apply and also that package
list can be generated and then we checked them. We should try to
generate and check for medkit as well.

Because someone on IRC (#turris on Freenode) complained that he could not
generate medkit.

It helped me to reproduce his issue with Gitlab CI and he didn't have installed pkg-config (dependency for Updater-ng, which is used to generate medkits), because it was working for me locally and also it works on our CI, but it does not provide any logs to public.
This can help users to know if medkits can be generated.