helpers/generate_medkit: fix missing BOARD
The BOARD variable was renamed to BOOTSTRAP_BOARD to be consistent with
bootstrap script in lists.

Another fix is to add actual support for BOOTSTRAP_BOARD in bootstrap
list.

Last modification is merge of now short file medkit-updater-ng.lua to
generate_medkit script. It can be now part of this script and thus
making this script self contained.