Add project specific gitconfig and mmore
This is a pull and modification of changes introduced to
turris-os-packages 848af2d4e4cd4c8591fbe8b525d303570e079a46.

It introduces primarily project specific config that includes new
scripts and hooks. To enable it in local repository it is just:
  git config --local include.path ../.gitconfig

The hook checks if commit being pushed to Gitlab server is correct.
There are multiple checks:
* Check naming convention
* Check base (hbk vs hbl vs hbd)
* Protects against forgotten hbl and hbd pushes when pushing hbk or hbl

In terms of aliases this adds:
git new-mr:
  This opens web browser with new merge request on https://gitlab.nic.cz
  for current branch.
git new-branch:
  This creates new branch interactively and should help users with
  naming convention.