feeds.conf: remove openwisp feed
We no longer need to have OpenWISP feed as OpenWISP was updated to the
latest version in these commits in branches: openwrt-19.07 [1] and master [2].

As it was updated, there are no differences and thus openwisp folder
(feed) on our repository is empty.

And if there are any differences, it should be added to OpenWrt's packages repository.

[1] https://github.com/openwrt/packages/blob/openwrt-19.07/admin/openwisp-config/Makefile
[2] https://github.com/openwrt/packages/commit/9f7b8088c374a5c8533c61baf56ae57343f6fb67