patches/openwrt: add fix for obligatory reported bugs in our packages
On feeds clone the script feeds generates indexes. It is doing that
right after it is cloned. That resulted in errors being reported for
some of our packages depending on packages feed. The most of those
packages were Python packages that were unable to be resolved thanks to
including file from packages feed that at that time was missing. This
just postpones index creation after all requested feeds are
updated/pulled.