generate_medkits: set default updater branch outside hbk
In default we do not configure updater branch thus using default of
updater-ng's package. That works for HBS and thus is needed to be same
for HBT and HBK. In case of other branches this is not exactly true and
we can simply use default branch as default update branch.