githooks: fix push reject when hbl is directly based on hbk
The issue this fixes is that git reports same commit as being ancestor
of itself. It kind of makes sense but it breaks check if branch is based
on hbk or hbl or hbd. In such case for example the common base between
hbl and target branch and hbk and target branch is the same.

This changes the logic. Because git merge-base does effectively the
operation of 'less than or equal' we have to swap compare order and
negate the result. This way we get effectively the 'less than'
operation.