configs/common/busybox: disable BUSYBOX_DEFAULT_INSTALL_NO_USR
In OpenWrt's defaults, this is disabled by default. Some scripts in our
distribution currently fails, because it could not find env.

Example:
/usr/bin/env is missing and it is very bad, because this is standard
path, where you expect env.

Because of this option, we can see that env is located in /bin, which is
wrong. We need this option in the past as we moved everything related to
busybox to /{s,}bin

More details about this option:
https://github.com/openwrt/openwrt/blob/bbff6239e2ea273388f4ca0f8586945ff5f36271/package/utils/busybox/config/Config.in#L195