Revert "patches/turrispackages: ruby-breakpoint: drop it"
This reverts commit ae63d8978ce5cac97b9c7186c8a6b52da7098858.
It seems that the transition to reForis does not make it soon (also Half
Life 3 is not coming anytime soon, right? ;)

Because OpenWrt 21.02 should be very soon released in RC1, we need to
revert this, because in that situation we will ship Foris and reForis
together until Foris is completely removed in the future. There are just
a few things which reForis misses - Nextcloud and Pakon