patches/openwrt: mvebu: initial kernel 5.10 support for Turris Omnia
LED driver for Turris Omnia was thanks to Marek added into kernel 5.10,
because of that, I removed it.

DSA multiCPU is not in kernel, yet.

Separate DTS, well, I guess, we need this, because there were changes in
upstream kernel, but it requires flashed newer U-boot, which we didn't
do.