patches/openwrt: tweak leds setup patch for Omnia
First change is that the LEDs should have been configured in leds
configuration file not in network one. This moves it there.

Second and most important change is the change from list of WiFi devices
to just plain wlan list for three wlan devices. This is in no way ideal
as it relies on initialization of PCI devices in order. This means that
leds won't no longer reflect exactly the position in PCI slot but rather
order of initialization.
The issue we are trying to solve here is that original code just wasn't
working as WiFi devices are not initialized at time of this script
execution. The result was that WiFi leds just were not configured at
all, at least till the system update that caused re-execution of this
script later in system execution.

And the last change is the move of this patch from branding to
to-upstream. We can't push it right now but we should be able in the
future once upstream updates kernel to version with our RGB leds
support.