lists/migrate3x: fix no access from LAN and ignored tos3to4-early
This fixes two issues.

The first issue is that updater is maked for replan in Turris OS 3.x
lists and it is replanned first becuase tos3to4-early depends on it. The
solution here is to replan one more time after tos3to4-early
installation that should be right after updater-ng is installed.

The second issue is that we migrate configuration from switch-config to
DSA but kernel we are running in that moment does not support DSA and
thus we end up with unconfigured LAN. This makes it hard to access
router after migration and makes router pretty much broken. The reboot
solves this as a new kernel is deployed. We do it this way without need
for user to reboot device manually.