lists/migrate3x: chain fix packages to tos3to4-early install
The tos3to4-early is implemented to migrate configuration as it as in
Turris OS 4.x. It makes no sense to duplicate script and logic that is
included in fix script, instead we can just simply use them. This chains
package list fix packages in a way they are executed after tos3to4-early
and thus they migrate result of tos3to4-early migration.