lists/base-fix: retrigger fix-dhparam-to-cagen
The fix was checking if file we plan to use is present just to not touch
it in case user placed it there. The issue is that in such case we won't
migrate the OpenVPN configuration. Our preference should be to migrate
the OpenVPN server configuration so we just removed this check and
release it as fixup.