lists/migrate3x: solve collision between cronie and vixie-cron
The tos3to4-early uses switch-branch that depends on updater-superviser
that in turn depends on cronie. We do not need updater-supervisor to
migrate updater's configuration and thus we can break this dependency
simply by temporally masking updater-supervisor behind virtual setting.
The updater-supervisor is considered as virtual until tos3to4-early is
not installed (here we use actually package we chain after tos3to4-early
to properly detect that). The last run that updates rest of the system
then updates updater-supervisor as usual.

The last change is just minor tweak. We add priority to tos3to4-early.
It is critically requested and thus priority plays no role in satisfying
requests but it is used to order requests and thus should force
tos3to4-early to be planned as soon as possible.