drivers/usb: remove RTL8812AU due to AP issues
Some users can be mistaken that any listed in this list is officially
supported, which is not because we don't have so many devices available
to test, and this list is mostly community-driven.

Based on the user, who reached our support department,  there was created GitLab
issue [1] that with the driver RTL8812AU, there are some issues with
OpenWrt, and they are discussed in the OpenWrt forum as well related to
AP mode.

There are some attemps to get alternative kernel drivers accepted to OpenWrt [2] [3].

Once they are accepted, we can add them back.

[1] https://gitlab.nic.cz/turris/os/build/-/issues/267
[2] https://github.com/openwrt/openwrt/pull/4229
[3] https://github.com/openwrt/openwrt/pull/4230