NEWS: more changes in Turris OS 5.2.4
On Friday, there was reported on our forum [1] that users can not download
LXC images anymore because it was using sks-keyservers.net as PGP
server, but this one is deprecated and no longer maintained.

That's why I updated package lxc everywhere to fix it.

Also, while at it, I updated mariadb.

[1] https://forum.turris.cz/t/lxc-create-error-unable-to-fetch-gpg-key-from-keyserver/15636