helpers/generate_common.sh: fix condition for default reference
The default reference should be used when no reference is given. The
original check was invalid because it checked that result is non-empty
but that never happens (ignoring case when someone used ';' or '^'
without specifying reference) as such URL misses ';' and '^'. The result
thus was whole URL.
This now correctly detects the case when reference is specified.