lists: immediately replan on updater-ng update
We need to replan immediately after updater-ng update because we depend
on conditions and without them a lot of packages are skipped and removed
during the update. We in general prefer to update whole system at once
but it is not an ideal when part of that system would be removed during
such update. It would also be better to use `Mode("no_removal")` but
that is available since version 64.0 and request conditions were
introduced in 65.0 thus it makes no sense to cover that single version
with special case. In the future when ever new incompatibility such as
this is added we should use no removal mode instead.