githooks/pre-push: allow fast-forward of hbk and hbl
The original check was that hbk and hbl can be pushed only with hbd and
hbl, in case of hbk, branches. It is what we mostly want but there are
cases when we do not update hbd or hbl and only hbk or hbl that are
valid as well. The example of such operation is merge of hbd to
hbl or merge of hbl to hbk, which is always fast-forward due to our
workflow.
This halts push only if hbk or hbl commit is not ancestor of hbd instead
of halting it always when hbd or hbl is not being updated.