helpers/new_release: fix with new lists git-hash format
The git-hash file for lists now has the same format as for packages.
Thus we have to test it separately as well as to extract hash from it
using the same function.
It is also now specified in feeds.conf file so we have to set
appropriate has to it as well.

This also adds `set -eu` to the script to make it more resilient.
There is also a tweak of indentation.