patches/openwrt/backport: backport new leds kmods
This backports one patch that is already in OpenWrt's master and another
one (mine) that is submitted to the master. I place it to to the
backport diretory as well as it is backport from master in the end.

These modules help to implement better rainbow. The uleds module can be
used to get LED triggers or even fake leds in a way to support old
platforms that do not have kernel driver.

The pattern trigger then can be used for animations.