patches/openwrt: fix invalid assignment on Shield (Mox)
The ucidef_set_interface_lan sets the one specific interface to LAN
network. If that suppose to be a bridge we have to provide all
interfaces at once. This changes the code so it does that.

One additional change here is that if there are no LAN ports the WAN
port is assigned to LAN instead of WAN.