NEWS: update changelog for Turris OS 5.3.4
As we didn't make this release last year, it does not make any sense to
wish Merry Christmas anymore. :-)

There was updated zsh and domoticz. The domoticz was requested on our
forum as it didnt work out with the Python 3.7. Both of them are
included or pending to by accepted in upstream