patches/openwrt: set full version for base-files
The OpenWrt uses revision as part of base-files and that works for them
but not for us. Our version is declared by turris-build repository and
applied as a patch. We also have patch that changes revision detection
to latest OpenWrt commit (ignoring our patches). That is done to not
reinstall base-files with pretty much every rebuild as hashes for
patches change with any change in OpenWrt. Thus it is easy for us to
have same version of base-files where the content of it is different.
The most problematic is /etc/os-release file.

The switch from revision to full version code causes reinstall of
package when we change our Turris OS version not just when OpenWrt
revision changes.