patches/openwrt: install initialization of fw_env.config as dedicated script
We sometimes need to update fw_env.config such as when we update u-boot
and want to migrate the environment. This installs previously only
temporally file as a dedicated script and thus to reset u-boot's
environment access tools's configuration is just in terms of removal of
the configuration file and calling this script.