patches/openwrt: u-boot-mvebu: update to version 2022.01
This patch series was submitted to OpenWrt [1], but it was somehow
agreed that new versions might cause regressions and that there is no
man power for testing it on each devices and the only way is for OpenWrt
release more often, but we can not force users to use the latest master
branches even though customers wants to have always the latest features
while using the stable branch.

We would like to sync all U-boot versions in upcoming Turris OS 6.x
releases, so we dont need to maintain various U-boot versions related to
DDR training on Marvell Armada 38X, we need to backport it and as
someone could be saying playing our playground instead of submitting
stuff to upstream, which we tried and we have at least reference.

For curious users, they can check following discussion on this thing on
IRC channel #openwrt-devel on OFTC server dated 21/3/2022.
(since 18:46:14, ...)

[1] https://github.com/openwrt/openwrt/pull/9409